pwa/backend/requirements.txt
django>=4.2
djangorestframework
django-cors-headers
django-filter
Pillow
python-decouple
